chore(migration): move types to @zenith/types
