+++ .gitignore
@@ -1,6 +1,8 @@
 .turbo
 .DS_Store
 node_modules
-# Local Netlify folder
-.netlify
+
+
 .vercel
+dist
+build
