+++ apps/spotify-backend/src/app/api/refresh/route.ts
@@ -2,7 +2,7 @@ import refreshAccessToken from "@/app/services/spotify-auth/refreshAccessToken";
 import { retriveAccessToken } from "@/app/services/token-storage";
 import { catchHttpErrors, throwHttpErrors } from "@/app/utils";
 import tokenStorageWorkflow from "@/app/workflow";
-import { IToken } from "@/types";
+import { IToken } from "@zenith/types";
 import { NextResponse } from "next/server";
 
 export const runtime = "edge";
