+++ apps/zenith-ui/src/vite-env.d.ts
@@ -1,9 +1,1 @@
 /// <reference types="vite/client" />
-
-interface IRoute {
-  path: string;
-  requireAuthentication: boolean;
-  allowAuthenticated?: false;
-  requiredRole?: string[];
-  element: JSX.Element;
-}
