+++ package.json
@@ -6,6 +6,7 @@
   "packageManager": "yarn@1.22.22",
   "scripts": {
     "build": "turbo build",
+    "build:types":"turbo run @zenith/types#build",
     "dev": "turbo dev",
     "start": "turbo start"
   },
@@ -14,6 +15,6 @@
     "packages/*"
   ],
   "devDependencies": {
-    "turbo": "^2.3.3"
+    "turbo": "^2.4.0"
   }
 }
