+++ packages/types/src/spotify/index.ts
@@ -1,0 +1,2 @@
+export * from "./token";
+export * from "./track";
