+++ packages/types/src/spotify/token.ts
@@ -1,0 +1,5 @@
+export interface IToken {
+  accessToken: string;
+  refreshToken: string;
+  expiresAt: number;
+}
