+++ packages/types/src/spotify/track.ts
@@ -1,0 +1,12 @@
+export interface ISpotifyImage {
+  url: string;
+  width: number;
+  height: number;
+}
+export interface ISpotifyTrackDetails {
+  name: string;
+  popularity?: number;
+  images: ISpotifyImage[];
+  uri: string;
+  artist: string;
+}
