+++ packages/types/src/ui/route.ts
@@ -1,0 +1,7 @@
+export interface IRoute {
+  path: string;
+  requireAuthentication: boolean;
+  allowAuthenticated?: false;
+  requiredRole?: string[];
+  element: React.ComponentType;
+}
