+++ turbo.json
@@ -2,6 +2,7 @@
   "$schema": "https://turbo.build/schema.json",
   "tasks": {
     "build": {
+      "dependsOn": ["^build"],
       "outputs": ["build/**", ".next/**", "dist/**"],
       "env": [
         "VITE_CLERK_PUBLISHABLE_KEY",
@@ -56,6 +57,10 @@
         "QSTASH_CURRENT_SIGNING_KEY",
         "QSTASH_NEXT_SIGNING_KEY"
       ]
+    },
+    "@zenith/types#build": {
+      "outputs": ["dist/**"],
+      "dependsOn": []
     }
   }
 }
